# Conflicts:
#	weasis-dicom/weasis-dicom-viewer2d/src/main/java/org/weasis/dicom/viewer2d/mpr/MprView.java
#	weasis-dicom/weasis-dicom-viewer2d/src/main/java/org/weasis/dicom/viewer2d/mpr/Volume.java
#	weasis-parent/pom.xml

+++ weasis-core/src/main/java/org/weasis/core/ui/tp/raven/datetime/component/time/PanelClock.java
@@ -208,7 +208,7 @@ public class PanelClock extends JPanel {
     }
     g2.setColor(getSelectedColor());
     float angle = animationChange.getAngle();
-    g2.rotate(Math.toRadians(angle), x + size, y + size);
+    g2.rotate(Math.toRadians(angle), (double) x + size, (double) y + size);
     g2.fill(area);
     g2.setTransform(tran);
   }
